# Conflicts:
#	out/production/MiningGame/game/Game.class
#	out/production/MiningGame/game/GameKeyListener.class
#	out/production/MiningGame/game/GameMap.class
#	out/production/MiningGame/game/GameScreen.class
#	out/production/MiningGame/game/objects/GameObject.class
#	out/production/MiningGame/game/objects/creatures/Player.class
#	out/production/MiningGame/game/objects/tiles/Air.class
#	out/production/MiningGame/game/objects/tiles/Block.class
#	out/production/MiningGame/game/objects/tiles/Tile.class

+++ game/GameMap.java
@@ -39,14 +39,14 @@ public class GameMap extends GameObject{
         {0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 0}
 	};
 
-	private final int tileSize;
+	private final double tileSize;
 	
 	private final Tile[][] tiles;
 	
 	
 	//Kontruktor
-	public GameMap(int tileSize) {
-		this.tileSize = (int) tileSize;
+	public GameMap(double tileSize) {
+		this.tileSize = tileSize;
 		tiles = new Tile[DEFAULT_MAP.length][DEFAULT_MAP[0].length];
 
 
@@ -66,7 +66,7 @@ public class GameMap extends GameObject{
 	
 	//Methoden
 	@Override
-	public void render(Graphics2D g, int tileSize) {
+	public void render(Graphics2D g, double tileSize) {
 		for (Tile[] row : tiles) {
 			for (Tile tile : row) {
 				tile.render(g, tileSize);
@@ -83,7 +83,7 @@ public class GameMap extends GameObject{
 		return tiles.length;
 	}
 	
-	public int getTileSize() {
+	public double getTileSize() {
 		return tileSize;
 	}
