+++ game/objects/tiles/Air.java
@@ -9,9 +9,8 @@ public class Air extends Tile{
 	}
 
 	@Override
-	public void render(Graphics2D g, int tileSize) {
-
-		//Air is invisible
+	public void render(Graphics2D g, double tileSize) {
+		//ignore
 	}
 
 }
